modeling/requirements.txt
argon2-cffi==20.1.0
astroid==2.5.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
decorator==5.0.6
defusedxml==0.7.1
entrypoints==0.3
greenlet==1.0.0
h5py==3.2.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupytext==1.11.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
llvmlite==0.36.0
markdown-it-py==0.6.2
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mdit-py-plugins==0.2.6
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numba==0.53.1
numpy==1.20.2
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
pycparser==2.20
Pygments==2.8.1
pylint==2.7.4
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.4.7
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.1
tornado==6.1
traitlets==5.0.5
tslearn==0.5.0.5
typing-extensions==3.10.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
